backend/requirements.txt
aiokafka==0.12.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
black==25.1.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
coverage==7.8.0
defusedxml==0.7.1
fastapi==0.115.12
flake8==7.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
kafka-python==2.1.5
mccabe==0.7.0
mypy==1.15.0
mypy_extensions==1.1.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
prometheus_client==0.21.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pyflakes==3.3.2
pyparsing==3.2.3
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
youtube-transcript-api==1.0.3
